[Nicolas Espina] Reparacion ingresada correctamente a traves de Postman
- Se guardan correctamente los valores de las reparaciones. falta ver la implementacion de los bonos.

- Se configura el frontEnd para el desarrollo del ingreso de las reparaciones.

+++ repairs_vehicle.ms/src/main/java/tingeso/ev2/repairs_vehicle/ms/entities/RepairVehicleEntity.java
@@ -36,7 +36,7 @@ public class RepairVehicleEntity {
     private double iva_mount;
     @Getter @Setter
     //mount with discounts, surcharges and bonus
-    private Long total_cost;
+    private double total_cost;
     @Getter @Setter
     private LocalDate vehicle_outcome;
     @Getter @Setter
